Doubles

+++ Core/Src/main.c
@@ -123,7 +123,7 @@ int main(void)
     /* USER CODE END WHILE */
 
     /* USER CODE BEGIN 3 */
-    LOGOMATIC("APPS1 is %lf\t\tAPPS2 is %lf\n", analogRead(APPS1_SIGNAL) * (long float)ADC_CONV, analogRead(APPS2_SIGNAL) * (long float)ADC_CONV);
+    LOGOMATIC("APPS1 is %lf\t\tAPPS2 is %lf\n", analogRead(APPS1_SIGNAL) * (double)ADC_CONV, analogRead(APPS2_SIGNAL) * (double)ADC_CONV);
 
     //writeMessage(PrimaryBusCAN, MSG_DEBUG_2_0, GR_ALL, (uint8_t*)"Testing", 8);
